BVH and Debug Renderer
~20hrs
- Fixed a bug with the vertex array binding offset calculation for the command list.
-- Did not account for joint allcoations by the allocator.
- New debug renderer with support for lines and Axis Aligned Boxes.
- Fixed a bug with the "BaseAssetHandle" inspector clear button where the id of the clear button was shared.
- Removing components are now more streamlined in the editor.
- Fix major bugs in the new Array C API where the old versions of functions were called casuing all sorts of memory errors.
- BVH tree implentation with rotaiosn for auto balancing and debug rendering.
- Models are now scaled to be relative to a standard1.0f for ease of bounding volume calculations.
- New Color serializers interfaces.
- const version of visit_all for Variant<Ts...>.
- ContainerTuple fixed for new version of MSVC as it  seemed it could not handle the old setup.
- General code cleanup + more comments in certain places.
- The camera init function can now accept nullptrs for inital position and it's world up.
-  Added some more functions to the templated math library.
- Thick line rendering initial implemtation, seems ok for now but needs some more work.

+++ BifrostScript/include/bifrost/script/bifrost_vm.h
@@ -257,7 +257,7 @@ struct BifrostVM_t
   struct BifrostObj_t*          temp_roots[8];                           /*!< Objects temporarily protected from the GC                                      */
   uint8_t                       temp_roots_top;                          /*!< BifrostVM_t::temp_roots size                                                   */
   bfBool32                      gc_is_running;                           /*!< This is so that when calling a finalizer the GC isn't run.                     */
-  size_t                        build_in_symbols[BIFROST_VM_SYMBOL_MAX]; /*!< Symbols that should be loaded at startup for a faster runtime.                 */
+  uint32_t                      build_in_symbols[BIFROST_VM_SYMBOL_MAX]; /*!< Symbols that should be loaded at startup for a faster runtime.                 */
   struct BifrostObjNativeFn_t*  current_native_fn;                       /*!< The currently executing native function for accessing of userdata and statics. */
 };
